1) TSS.JS/Java/Py: Removed unnecessary constatnt tag member from tagged structures. 2) TSS.JS/Java/Py: Replaced manual member-by-member initialization of the command structure with initialization list. 3) TSS.Java: Eliminated empty handle lists, and fixed pretty-printed types of some data structure members. 4) TSS.JS: Eliminated empty response data structures (replaced them with just 'void'). 5) TSS.Net: Eliminated elementary and duplicate types in the [KnownType()] class attributes.

+++ TSS.Java/src/tss/tpm/CommitResponse.java
@@ -19,12 +19,12 @@ public class CommitResponse extends TpmStructure
      * @param _E ECC point E [r]P1 
      * @param _counter least-significant 16 bits of commitCount
      */
-    public CommitResponse(TPMS_ECC_POINT _K,TPMS_ECC_POINT _L,TPMS_ECC_POINT _E,int _counter)
+    public CommitResponse(TPMS_ECC_POINT _K,TPMS_ECC_POINT _L,TPMS_ECC_POINT _E,short _counter)
     {
         K = _K;
         L = _L;
         E = _E;
-        counter = (short)_counter;
+        counter = _counter;
     }
     /**
     * TPM2_Commit() performs the first part of an ECC anonymous signing operation. The TPM will perform the point multiplications on the provided points and return intermediate signing values. The signHandle parameter shall refer to an ECC key and the signing scheme must be anonymous (TPM_RC_SCHEME).
