+++ TSS.Java/src/tss/tpm/TPM2_PolicyCounterTimer_REQUEST.java
@@ -19,11 +19,11 @@ public class TPM2_PolicyCounterTimer_REQUEST extends TpmStructure
      * @param _offset the octet offset in the TPMS_TIME_INFO structure for the start of operand A 
      * @param _operation the comparison to make
      */
-    public TPM2_PolicyCounterTimer_REQUEST(TPM_HANDLE _policySession,byte[] _operandB,int _offset,TPM_EO _operation)
+    public TPM2_PolicyCounterTimer_REQUEST(TPM_HANDLE _policySession,byte[] _operandB,short _offset,TPM_EO _operation)
     {
         policySession = _policySession;
         operandB = _operandB;
-        offset = (short)_offset;
+        offset = _offset;
         operation = _operation;
     }
     /**
