+++ TSS.Java/src/tss/tpm/TPMS_ALGORITHM_DETAIL_ECC.java
@@ -26,10 +26,10 @@ public class TPMS_ALGORITHM_DETAIL_ECC extends TpmStructure
      * @param _n order of G 
      * @param _h cofactor (a size of zero indicates a cofactor of 1)
      */
-    public TPMS_ALGORITHM_DETAIL_ECC(TPM_ECC_CURVE _curveID,int _keySize,TPMU_KDF_SCHEME _kdf,TPMU_ASYM_SCHEME _sign,byte[] _p,byte[] _a,byte[] _b,byte[] _gX,byte[] _gY,byte[] _n,byte[] _h)
+    public TPMS_ALGORITHM_DETAIL_ECC(TPM_ECC_CURVE _curveID,short _keySize,TPMU_KDF_SCHEME _kdf,TPMU_ASYM_SCHEME _sign,byte[] _p,byte[] _a,byte[] _b,byte[] _gX,byte[] _gY,byte[] _n,byte[] _h)
     {
         curveID = _curveID;
-        keySize = (short)_keySize;
+        keySize = _keySize;
         kdf = _kdf;
         sign = _sign;
         p = _p;
